改为函数;  检测是否有 项目nodejs环境激活文件

+++ nodejs_script/create-vite-wrap.sh
@@ -15,24 +15,32 @@ set -e -u
 
 source /app/bash-simplify/argCntEq1.sh
 
+function create_vite_wrap(){
 
 # 若函数参数不为1个 ， 则 返回错误
 argCntEq1 $* || { return $exitCode ;}
 
  _PrjHome=$1
 # _PrjHome=/app2/ncre
+
+#检测是否有 项目nodejs环境激活文件 PrjNodeJsEnvActivate.sh
+_prjNodeJsEnvActv_F=$_PrjHome/PrjNodeJsEnvActivate.sh
+_err2Msg="错误, 无nodejs环境激活文件[$_prjNodeJsEnvActv_F], 请先以此创建nodejs项目环境[/app/bash-simplify/nodejs_script/new_nodejsPrj_by_nodeenv.sh]"
+_err2=2
+[[ -f $_prjNodeJsEnvActv_F ]] || { echo $_err2Msg ; return $_err2 ;}
+
 _PrjName=$(basename $_PrjHome)
 
 _tmpHome=$_PrjHome/.tmp
 _tmpPrjHome=$_tmpHome/$_PrjName
 
-mkdir $_tmpHome; cd $_tmpHome
+rm -fr $_tmpHome; mkdir $_tmpHome; cd $_tmpHome
 
 #创建项目
 # create-vite --template vue
 create-vite
 
-_err1Msg="错误, 请务必输入同名项目名, 项目名必须为[$_PrjName], 已删除目录[$_tmpPrjHome]"
+_err1Msg="错误, 项目名必须同名为[$_PrjName], 已删除目录[$_tmpHome]"
 _err1=1
 [[ -d $_tmpPrjHome ]] || { echo $_err1Msg; rm -fr $_tmpPrjHome; return $_err1 ;}
 
@@ -43,3 +51,6 @@ mv --verbose $_tmpHome/* $_PrjHome/
 #删除临时目录
 rm -fr $_tmpHome
 
+}
+
+create_vite_wrap $*
